Show disconnect error

+++ FileSharing/clientfilethread.cpp
@@ -74,12 +74,12 @@ void ClientFileThread::downloadFile()
 	qDebug()<<fileSize;
 	while(fileSize > 0)
 	{
-        /*if (!peer->waitForReadyRead(-1)){
+        if (peer->state() != QAbstractSocket::ConnectedState){
             file.close();
             QFile::remove(downloadPath);
             emit connectionFailed(downloadPath);
             return;
-        }*/
+        }
         data = peer->readAll();
 		received = data.count();
 		sum_received += received;
